+++ example/tower/tower_base_greebled.py
@@ -37,6 +37,12 @@ bp_tower_base.door_width = 27
 bp_tower_base.door_height = 50
 bp_tower_base.door_padding = 7
 bp_tower_base.door_count = 1
+
+bp_tower_base.render_magnets = True
+bp_tower_base.magnet_diameter = 3.4
+bp_tower_base.magnet_height = 2.2
+bp_tower_base.magnet_count = 4
+
 bp_tower_base.make()
 ex_tower = bp_tower_base.build()
 #show_object(ex_tower)