+++ src/cqfantasy/tower/__init__.py
@@ -1,5 +1,6 @@
 
 from .cut_cylinder import cut_cylinder
+from .magnets import make_magnet, make_magnets
 from .TowerWindow import TowerWindow
 from .FrameWindow import FrameWindow
 from .LatticeWindow import LatticeWindow